+++ toolchain/sem_ir/BUILD
@@ -36,6 +36,7 @@ cc_library(
         "//common:check",
         "//common:enum_base",
         "//common:ostream",
+        "//common:type_enum",
         "//toolchain/base:index_base",
         "//toolchain/base:int",
         "//toolchain/base:value_ids",
@@ -162,6 +163,7 @@ cc_library(
         "//common:concepts",
         "//common:ostream",
         "//common:raw_string_ostream",
+        "//common:type_enum",
         "//toolchain/base:kind_switch",
         "//toolchain/base:shared_value_stores",
         "//toolchain/base:value_ids",
